Revert "Split out separate unit tests"
This reverts commit 31681468ab09aec698f824c4b24b5eaec96540cb.
